Commit 2
Implemented Adapter between the controller and the soon to be supplied Filehandler. This is because it is expected that the return types of the file handler will not use the Level object i am using. It is likely that i will need to use different filehandlers plugged into the Filehandler interface anyway so being able to implement different adapters to different filehandlers will be easier.